Major project/requirements.txt
Flask>=2.0
cryptography>=3.4
joblib>=1.0
numpy>=1.20
scikit-learn>=1.0
requests>=2.25
mysql-connector-python>=8.0
Pillow>=8.0
python-dotenv>=0.21
